Client validation

+++ flutter_todo/lib/components/todo_item.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_todo/components/modify_item.dart';
+import 'package:flutter_todo/components/widgets.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_todo/realm/schemas.dart';
 import 'package:flutter_todo/realm/realm_services.dart';
@@ -14,31 +15,41 @@ class TodoItem extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final realmServices = Provider.of<RealmServices>(context);
+    bool isMine = (item.ownerId == realmServices.currentUser?.id);
     return item.isValid
         ? ListTile(
-      leading: Checkbox(
-        checkColor: Colors.white,
-        fillColor: MaterialStateProperty.resolveWith(getColor),
-        value: item.isComplete,
-        onChanged: (bool? value) async {
-          await realmServices.updateItem(item, isComplete: value ?? false);
-        },
-      ),
+            leading: Checkbox(
+              checkColor: Colors.white,
+              fillColor: MaterialStateProperty.resolveWith(getColor),
+              value: item.isComplete,
+              onChanged: (bool? value) async {
+                await realmServices.updateItem(item, isComplete: value ?? false);
+              },
+            ),
             title: Text(item.summary),
-            subtitle: Text("${item.isComplete ? 'Completed' : 'Incomplete'} ${item.ownerId == realmServices.currentUser?.id ? ' - mine' : ''}"),
-      trailing: PopupMenuButton<MenuOption>(
-        onSelected: (menuItem) => handleMenuClick(context, menuItem, item, realmServices),
-        itemBuilder: (context) => [
-          const PopupMenuItem<MenuOption>(
-            value: MenuOption.edit,
-            child: ListTile(leading: Icon(Icons.edit), title: Text("Edit item")),
-          ),
-          const PopupMenuItem<MenuOption>(
-            value: MenuOption.delete,
-            child: ListTile(leading: Icon(Icons.delete), title: Text("Delete item")),
-          ),
-        ],
-      ),
+            subtitle: Row(
+              children: [
+                Text(isMine ? '(mine) ' : '', style: const TextStyle(fontWeight: FontWeight.bold, color: Colors.black)),
+                Text(item.isComplete ? 'Completed' : 'Incomplete'),
+              ],
+            ),
+            trailing: SizedBox(
+              width: 25,
+              child: PopupMenuButton<MenuOption>(
+                onSelected: (menuItem) => handleMenuClick(context, menuItem, item, realmServices),
+                itemBuilder: (context) => [
+                  const PopupMenuItem<MenuOption>(
+                    value: MenuOption.edit,
+                    child: ListTile(leading: Icon(Icons.edit), title: Text("Edit item")),
+                  ),
+                  const PopupMenuItem<MenuOption>(
+                    value: MenuOption.delete,
+                    child: ListTile(leading: Icon(Icons.delete), title: Text("Delete item")),
+                  ),
+                ],
+              ),
+            ),
+            shape: const Border(bottom: BorderSide()),
           )
         : Container();
   }
@@ -46,16 +57,37 @@ class TodoItem extends StatelessWidget {
   void handleMenuClick(BuildContext context, MenuOption menuItem, Item item, RealmServices realmServices) {
     switch (menuItem) {
       case MenuOption.edit:
-        showModalBottomSheet(
-          context: context,
-          isScrollControlled: true,
-          builder: (_) => Wrap(children: [ModifyItemForm(item)]),
-        );
+        if (item.ownerId == realmServices.currentUser?.id) {
+          showModalBottomSheet(
+            context: context,
+            isScrollControlled: true,
+            builder: (_) => Wrap(children: [ModifyItemForm(item)]),
+          );
+        } else {
+          showError(context, '''You are not allowed to edit tasks
+            that don't belogs to you.''');
+        }
         break;
       case MenuOption.delete:
-        realmServices.deleteItem(item);
+        if (item.ownerId == realmServices.currentUser?.id) {
+          realmServices.deleteItem(item);
+        } else {
+          showError(context, '''You are not allowed to delete tasks
+            that don't belogs to you.''');
+        }
+
         break;
     }
+  }
+
+  void showError(BuildContext context, String error, {int durationInSeconds = 15}) {
+    ScaffoldMessenger.of(context).hideCurrentSnackBar();
+    ScaffoldMessenger.of(context).showSnackBar(
+      errorMessageWidget(error),
+    );
+    Future.delayed(Duration(seconds: durationInSeconds)).then((value) {
+      ScaffoldMessenger.of(context).hideCurrentSnackBar();
+    });
   }
 
   Color getColor(Set<MaterialState> states) {
